extracted custom hook and utils to slim down appointments list component
